Lab2: reorganize resource panel components
