+++ apps/src/lab2/views/components/Instructions/ResourcePanel/AiTutorChatWithInstructionsDrawer/ai-tutor-chat-with-instructions-drawer.module.scss
