+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Footer/SettingsPanel.tsx
