+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Footer/disclaimer.module.scss
