+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Footer/settings-panel.module.scss
