+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Validation/ValidationPanel.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 
-import ValidationButton from '../ValidationButton';
-
+import ValidationButton from '../../ValidationButton';
 import {
   resourcePanelValidationTableElementId,
   resourcePanelValidateButtonElementId,
-} from './constants';
+} from '../constants';
+
 import ValidationTable from './ValidationTable';
 
 import validationStyles from './validation-panel.module.scss';
