+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Validation/ValidationTable.tsx
