+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Validation/ValidationTourSteps.tsx
@@ -11,8 +11,9 @@ import {
   RESOURCE_PANEL_PINNED_BUTTON_ONBOARDING_TOUR_SEEN,
   resourcePanelTabValidationElementId,
   resourcePanelValidateButtonElementId,
-} from './constants';
-import {Tabs} from './types';
+} from '../constants';
+import {Tabs} from '../types';
+
 import {VALIDATION_TOUR_STEPS} from './validationTourHelpers';
 
 // Check if tour should be disabled (e.g., during UI tests) before any rendering.
