+++ apps/src/lab2/views/components/Instructions/ResourcePanel/Validation/validation-panel.module.scss
